quiz - 자가 호출로 인한 Transaction 적용 불가 -> 실제 DB 호출용 클래스 분리
